separate voicing-dictionary & voice-leading
