+++ packages/voice-leading/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@tonaljs/voicing",
+  "name": "@tonaljs/voice-leading",
   "version": "4.5.1",
-  "description": "Voicings and Voice Leading for Chords",
+  "description": "Voice leading logic for transitions between voicings",
   "keywords": [
     "chord",
     "voicing",
@@ -17,12 +17,7 @@
   ],
   "types": "dist/index.d.ts",
   "dependencies": {
-    "@tonaljs/chord": "^4.5.1",
-    "@tonaljs/note": "^4.5.1",
-    "@tonaljs/range": "^4.5.1",
-    "@tonaljs/interval": "^4.5.1",
-    "@tonaljs/voicing-dictionary": "^4.5.1",
-    "@tonaljs/voice-leading": "^4.5.1"
+    "@tonaljs/note": "^4.5.1"
   },
   "author": "felix91@gmail.com",
   "license": "MIT",
